refactor keywords
