+++ src/main/java/org/kish/KishServer.java
@@ -1,6 +1,6 @@
-package org.kish2020;
+package org.kish;
 
-import org.kish2020.dataBase.ExpandedDataBase;
+import org.kish.dataBase.ExpandedDataBase;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.event.EventListener;
@@ -10,7 +10,7 @@ import java.util.Date;
 import java.util.HashMap;
 
 @SpringBootApplication
-public class Kish2020Server {
+public class KishServer {
     public static ExpandedDataBase mainDataBase = null;
     public static ExpandedDataBase mainSettings = null;
     public static FirebaseManager firebaseManager = null;
@@ -20,7 +20,7 @@ public class Kish2020Server {
         mainSettings = new ExpandedDataBase("db/kish2020.json");
         mainSettings.setSaveWithPrettyGson(true);
         firebaseManager = new FirebaseManager();
-        SpringApplication.run(Kish2020Server.class, args);
+        SpringApplication.run(KishServer.class, args);
 
         firebaseManager.sendFCMToAdmin("Server started", "서버가 시작되었습니다.\n" + new Date().toString(), new HashMap<>());
     }
