+++ src/main/java/org/kish/dataBase/DataBase.java
@@ -1,4 +1,4 @@
-package org.kish2020.dataBase;
+package org.kish.dataBase;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -6,7 +6,7 @@ import org.apache.commons.io.FileUtils;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
-import org.kish2020.MainLogger;
+import org.kish.MainLogger;
 
 import java.io.File;
 import java.io.IOException;
